requirements.txt
requests
discord.py
pyyamlconfig
aiohttp
